+++ app/src/models/cv.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct CV {
-    pub id: i32,
-    pub user_id: i32,
+    pub id: usize,
+    pub user_id: usize,
     pub cv_name: String,
     pub file_name: String,
-    pub keyword_list: Vec<i32>,
-    pub target_companies: Vec<i32>,
-    pub target_job_functions: Vec<i32>,
+    pub keyword_list: Vec<usize>,
+    pub target_companies: Vec<usize>,
+    pub target_job_functions: Vec<usize>,
     pub date_created: NaiveDateTime,
 }
 
@@ -17,10 +17,10 @@ impl CV {}
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct NewCV {
-    pub user_id: i32,
+    pub user_id: usize,
     pub cv_name: String,
     pub file_name: String,
-    pub keyword_list: Vec<i32>,
-    pub target_companies: Vec<i32>,
-    pub target_job_functions: Vec<i32>,
+    pub keyword_list: Vec<usize>,
+    pub target_companies: Vec<usize>,
+    pub target_job_functions: Vec<usize>,
 }
