+++ app/src/models/job_function.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct JobFunction {
-    pub id: i32,
+    pub id: usize,
     pub job_function_name: String,
-    pub keyword_list: Vec<i32>,
+    pub keyword_list: Vec<usize>,
     pub date_created: NaiveDateTime,
 }
 
@@ -14,5 +14,5 @@ impl JobFunction {}
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct NewJobFunction {
     pub job_function_name: String,
-    pub keyword_list: Vec<i32>,
+    pub keyword_list: Vec<usize>,
 }
